REGRESSION(r258493): CRASH running fast/mediastream/RTCPeerConnection tests https://bugs.webkit.org/show_bug.cgi?id=211666 <rdar://problem/63055644>
Reviewed by Eric Carlson.

MockMediaStreamTrack advertises that it's a webrtc::VideoTrackInterface, but doesn't inherit from
that interface, leading to a crash when its pointer is cast to webrtc::VideoTrackInterface and a
virtual method called on the resulting pointer. Since it's only used in one place, and since there's
an existing mock webrtc::VideoTrackInterface object in that same file, remove MockMediaStreamTrack
and replace the one place it's used with MockLibWebRTCVideoTrack.

* testing/MockLibWebRTCPeerConnection.h:
(WebCore::MockMediaStreamTrack::state const): Deleted.


Identifier: 224593@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@261447 268f45cc-cd09-0410-ab3c-d52691b4dbfc