Wasm::enableFastMemory() was called too late. https://bugs.webkit.org/show_bug.cgi?id=211773
Reviewed by Yusuke Suzuki.

Source/JavaScriptCore:

If Wasm fast memory is to be enabled, we should just do it in initializeThreading()
just like for all the other signal handlers that need to be initialized for JSC.
This simplifies its initialization and ensures that it is done in a timely manner
before Configs are frozen.

* jsc.cpp:
(jscmain):
* runtime/InitializeThreading.cpp:
(JSC::initializeThreading):

Source/WebKit:

Previously, we were calling Wasm::enableFastMemory() in WebProcess::initializeWebProcess().
By then, it is too late because the JSC::Config and WTF::Config have already been
frozen.  There is nothing special about Wasm::enableFastMemory() that it needs to
be called this late.  This patch moves it into JSC::initializeThreading() where
it belongs.

This change also make the initialization of Wasm fast memory consistent with how
we have always done it for the jsc shell i.e. it is initialized at the end of
JSC::initializeThreading().

* WebProcess/WebProcess.cpp:
(WebKit::WebProcess::initializeWebProcess):



Identifier: 224667@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@261549 268f45cc-cd09-0410-ab3c-d52691b4dbfc