Check for accessibilityEnabled() before posting notifications asynchronously. https://bugs.webkit.org/show_bug.cgi?id=211848
Reviewed by Chris Fleizach.

Covered by multiple tests. Fixes crashes in LayoutTests in isolated tree mode.

During LayoutTests, accessibility may be disabled between the time the
notifications are queued and the timer fires. Thus it is necessary to
check for accessibilityEnabled() before posting the notifications
asynchronously. Not doing so was causing crashes in isolated mode.

* accessibility/AXObjectCache.cpp:
(WebCore::AXObjectCache::notificationPostTimerFired):


Identifier: 224731@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@261625 268f45cc-cd09-0410-ab3c-d52691b4dbfc