Allow clipboard API access when pasting from a menu item or key binding https://bugs.webkit.org/show_bug.cgi?id=211990 <rdar://problem/63308916>
Reviewed by Megan Gardner.

Source/WebCore:

Allow the contents of the clipboard to be programmatically requested by the page while pasting from trusted UI
(i.e. the paste menu item, or when WebKit API is called by the app to trigger the paste). This allows the
'reading' part of the async clipboard API (`read` and `readText`) to be used when the user pastes in an editable
element, without having to fall back to showing the DOM paste access menu.

Note that this change should not have an effect on the pasteboard security model, since it only grants the page
programmatic access to the contents of the pasteboard in the case where access to the pasteboard has already
been granted by the user. Additionally, even in the event that the web process is compromised, even if the web
process can be tricked into believing it has been granted pasteboard access, the changes in r259151 will prevent
it from being able to request pasteboard data, unless the user (or the application, on behalf of the user) has
explicitly pasted via trusted API calls that are inaccessible from the web process.

Test: editing/async-clipboard/clipboard-read-while-pasting.html

* editing/Editor.cpp:
(WebCore::Editor::paste):
(WebCore::Editor::pasteAsPlainText):
(WebCore::Editor::pasteAsQuotation):

If `FromMenuOrKeyBinding::Yes` is passed in, set the `m_pastingFromMenuOrKeyBinding` flag to true during the
scope of the paste command.

* editing/Editor.h:
(WebCore::Editor::isPastingFromMenuOrKeyBinding const):
* editing/EditorCommand.cpp:
(WebCore::executePaste):
(WebCore::executePasteAndMatchStyle):
(WebCore::executePasteAsPlainText):
(WebCore::executePasteAsQuotation):

Pass in `FromMenuOrKeyBinding::Yes` when triggering the paste from a menu item or key binding.

* page/Frame.cpp:
(WebCore::Frame::requestDOMPasteAccess):

When pasting from menu or key binding, grant the page DOM paste access without requiring the DOM paste access
UI to be shown and confirmed.

Tools:

Add a UIScriptController method to trigger a paste from the application process.

* DumpRenderTree/cocoa/UIScriptControllerCocoa.h:
* DumpRenderTree/cocoa/UIScriptControllerCocoa.mm:
(WTR::UIScriptControllerCocoa::paste):
* TestRunnerShared/UIScriptContext/Bindings/UIScriptController.idl:
* TestRunnerShared/UIScriptContext/UIScriptController.h:
(WTR::UIScriptController::paste):
* WebKitTestRunner/cocoa/UIScriptControllerCocoa.h:
* WebKitTestRunner/cocoa/UIScriptControllerCocoa.mm:
(WTR::UIScriptControllerCocoa::paste):

LayoutTests:

Add a new layout test to verify that the contents of the clipboard can be read while performing a paste that was
not triggered from the DOM.

* editing/async-clipboard/clipboard-read-while-pasting-expected.txt: Added.
* editing/async-clipboard/clipboard-read-while-pasting.html: Added.
* platform/win/TestExpectations:
* resources/ui-helper.js:
(window.UIHelper.async copyText):
(window.UIHelper.async paste):


Identifier: 224916@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@261825 268f45cc-cd09-0410-ab3c-d52691b4dbfc