REGRESSION (r258919): fast/events/ios/key-events-comprehensive/key-events-control tests are failing https://bugs.webkit.org/show_bug.cgi?id=212124 <rdar://problem/60893744>
Reviewed by Tim Horton.

After the changes in r258919, WKContentView's text input traits report that we're in a single line input, even
when there is no editable focused element. This prevents UIKit from allowing WebKit to handle the `CTRL + M`
key command, which is interpreted at the system level as inserting a line break in a multi-line editable area.

It seems odd to consider ourselves in a single line input when there isn't even an editable element focused
element; for now, we can fix these failing tests by just going with the default value for the text input traits'
`setIsSingleLineDocument` property, but not explicitly setting it to either `NO` or `YES` when no element is
focused.

* UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView textInputTraits]):


Identifier: 224992@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@261906 268f45cc-cd09-0410-ab3c-d52691b4dbfc